+++ R/lambert.R
@@ -24,6 +24,10 @@
 #'   this estimate 1-1 transform. These alternative types are effective when the
 #'   data has exceptionally heavy tails, e.g. the Cauchy distribution.
 #'   
+#'   Additionally, sometimes (depending on the distribution) this method will be
+#'   unable to extrapolate beyond the observed bounds. In this cases, NaN is
+#'   returned
+#'   
 #' @return A list of class \code{lambert} with elements 
 #'   \item{x.t}{transformed original data} 
 #'   \item{x}{original data}